refactor(commands): extract registry and parser helpers
